requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
pydantic[email]==2.7.1
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0
pytest-cov==4.1.0
